+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc/transaction.ex
@@ -11,6 +11,39 @@ defmodule EthereumJSONRPC.Transaction do
 
   alias EthereumJSONRPC
 
+  case Application.compile_env(:explorer, :chain_type) do
+    "ethereum" ->
+      @chain_type_fields quote(
+                           do: [
+                             max_fee_per_blob_gas: non_neg_integer(),
+                             blob_versioned_hashes: [EthereumJSONRPC.hash()]
+                           ]
+                         )
+
+    "suave" ->
+      @chain_type_fields quote(
+                           do: [
+                             execution_node_hash: EthereumJSONRPC.address(),
+                             wrapped_type: non_neg_integer(),
+                             wrapped_nonce: non_neg_integer(),
+                             wrapped_to_address_hash: EthereumJSONRPC.address(),
+                             wrapped_gas: non_neg_integer(),
+                             wrapped_gas_price: non_neg_integer(),
+                             wrapped_max_priority_fee_per_gas: non_neg_integer(),
+                             wrapped_max_fee_per_gas: non_neg_integer(),
+                             wrapped_value: non_neg_integer(),
+                             wrapped_input: String.t(),
+                             wrapped_v: non_neg_integer(),
+                             wrapped_r: non_neg_integer(),
+                             wrapped_s: non_neg_integer(),
+                             wrapped_hash: EthereumJSONRPC.hash()
+                           ]
+                         )
+
+    _ ->
+      @chain_type_fields quote(do: [])
+  end
+
   @type elixir :: %{
           String.t() => EthereumJSONRPC.address() | EthereumJSONRPC.hash() | String.t() | non_neg_integer() | nil
         }
@@ -42,8 +75,17 @@ defmodule EthereumJSONRPC.Transaction do
    * `"maxPriorityFeePerGas"` - `t:EthereumJSONRPC.quantity/0` of wei to denote max priority fee per unit of gas used. Introduced in [EIP-1559](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md)
    * `"maxFeePerGas"` - `t:EthereumJSONRPC.quantity/0` of wei to denote max fee per unit of gas used. Introduced in [EIP-1559](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md)
    * `"type"` - `t:EthereumJSONRPC.quantity/0` denotes transaction type. Introduced in [EIP-1559](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md)
-   * `"executionNode"` - `t:EthereumJSONRPC.address/0` of execution node (used by Suave).
-   * `"requestRecord"` - map of wrapped transaction data (used by Suave).
+   #{case Application.compile_env(:explorer, :chain_type) do
+    "ethereum" -> """
+       * `"maxFeePerBlobGas"` - `t:EthereumJSONRPC.quantity/0` of wei to denote max fee per unit of blob gas used. Introduced in [EIP-4844](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-4844.md)
+       * `"blobVersionedHashes"` - `t:list/0` of `t:EthereumJSONRPC.hash/0` of included data blobs hashes. Introduced in [EIP-4844](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-4844.md)
+      """
+    "suave" -> """
+       * `"executionNode"` - `t:EthereumJSONRPC.address/0` of execution node (used by Suave).
+       * `"requestRecord"` - map of wrapped transaction data (used by Suave).
+      """
+    _ -> ""
+  end}
   """
   @type t :: %{
           String.t() =>
@@ -51,6 +93,7 @@ defmodule EthereumJSONRPC.Transaction do
         }
 
   @type params :: %{
+          unquote_splicing(@chain_type_fields),
           block_hash: EthereumJSONRPC.hash(),
           block_number: non_neg_integer(),
           from_address_hash: EthereumJSONRPC.address(),
@@ -68,21 +111,7 @@ defmodule EthereumJSONRPC.Transaction do
           transaction_index: non_neg_integer(),
           max_priority_fee_per_gas: non_neg_integer(),
           max_fee_per_gas: non_neg_integer(),
-          type: non_neg_integer(),
-          execution_node_hash: EthereumJSONRPC.address(),
-          wrapped_type: non_neg_integer(),
-          wrapped_nonce: non_neg_integer(),
-          wrapped_to_address_hash: EthereumJSONRPC.address(),
-          wrapped_gas: non_neg_integer(),
-          wrapped_gas_price: non_neg_integer(),
-          wrapped_max_priority_fee_per_gas: non_neg_integer(),
-          wrapped_max_fee_per_gas: non_neg_integer(),
-          wrapped_value: non_neg_integer(),
-          wrapped_input: String.t(),
-          wrapped_v: non_neg_integer(),
-          wrapped_r: non_neg_integer(),
-          wrapped_s: non_neg_integer(),
-          wrapped_hash: EthereumJSONRPC.hash()
+          type: non_neg_integer()
         }
 
   @doc """
@@ -168,82 +197,13 @@ defmodule EthereumJSONRPC.Transaction do
       }
   """
   @spec elixir_to_params(elixir) :: params
-
-  # this is for Suave chain (handles `executionNode` and `requestRecord` fields along with EIP-1559 fields)
-  def elixir_to_params(
-        %{
-          "blockHash" => block_hash,
-          "blockNumber" => block_number,
-          "from" => from_address_hash,
-          "gas" => gas,
-          "gasPrice" => gas_price,
-          "hash" => hash,
-          "input" => input,
-          "nonce" => nonce,
-          "r" => r,
-          "s" => s,
-          "to" => to_address_hash,
-          "transactionIndex" => index,
-          "v" => v,
-          "value" => value,
-          "type" => type,
-          "maxPriorityFeePerGas" => max_priority_fee_per_gas,
-          "maxFeePerGas" => max_fee_per_gas,
-          "executionNode" => execution_node_hash,
-          "requestRecord" => wrapped
-        } = transaction
-      ) do
-    result = %{
-      block_hash: block_hash,
-      block_number: block_number,
-      from_address_hash: from_address_hash,
-      gas: gas,
-      gas_price: gas_price,
-      hash: hash,
-      index: index,
-      input: input,
-      nonce: nonce,
-      r: r,
-      s: s,
-      to_address_hash: to_address_hash,
-      v: v,
-      value: value,
-      transaction_index: index,
-      type: type,
-      max_priority_fee_per_gas: max_priority_fee_per_gas,
-      max_fee_per_gas: max_fee_per_gas
-    }
-
-    # credo:disable-for-next-line
-    result =
-      if Application.get_env(:explorer, :chain_type) == "suave" do
-        Map.merge(result, %{
-          execution_node_hash: execution_node_hash,
-          wrapped_type: quantity_to_integer(Map.get(wrapped, "type")),
-          wrapped_nonce: quantity_to_integer(Map.get(wrapped, "nonce")),
-          wrapped_to_address_hash: Map.get(wrapped, "to"),
-          wrapped_gas: quantity_to_integer(Map.get(wrapped, "gas")),
-          wrapped_gas_price: quantity_to_integer(Map.get(wrapped, "gasPrice")),
-          wrapped_max_priority_fee_per_gas: quantity_to_integer(Map.get(wrapped, "maxPriorityFeePerGas")),
-          wrapped_max_fee_per_gas: quantity_to_integer(Map.get(wrapped, "maxFeePerGas")),
-          wrapped_value: quantity_to_integer(Map.get(wrapped, "value")),
-          wrapped_input: Map.get(wrapped, "input"),
-          wrapped_v: quantity_to_integer(Map.get(wrapped, "v")),
-          wrapped_r: quantity_to_integer(Map.get(wrapped, "r")),
-          wrapped_s: quantity_to_integer(Map.get(wrapped, "s")),
-          wrapped_hash: Map.get(wrapped, "hash")
-        })
-      else
-        result
-      end
-
-    put_if_present(transaction, result, [
-      {"creates", :created_contract_address_hash},
-      {"block_timestamp", :block_timestamp}
-    ])
+  def elixir_to_params(elixir) do
+    elixir
+    |> do_elixir_to_params()
+    |> chain_type_fields(elixir)
   end
 
-  def elixir_to_params(
+  def do_elixir_to_params(
         %{
           "blockHash" => block_hash,
           "blockNumber" => block_number,
@@ -293,7 +253,7 @@ defmodule EthereumJSONRPC.Transaction do
 
   # txpool_content method on Erigon node returns tx data
   # without gas price
-  def elixir_to_params(
+  def do_elixir_to_params(
         %{
           "blockHash" => block_hash,
           "blockNumber" => block_number,
@@ -340,77 +300,7 @@ defmodule EthereumJSONRPC.Transaction do
     ])
   end
 
-  # this is for Suave chain (handles `executionNode` and `requestRecord` fields without EIP-1559 fields)
-  def elixir_to_params(
-        %{
-          "blockHash" => block_hash,
-          "blockNumber" => block_number,
-          "from" => from_address_hash,
-          "gas" => gas,
-          "gasPrice" => gas_price,
-          "hash" => hash,
-          "input" => input,
-          "nonce" => nonce,
-          "r" => r,
-          "s" => s,
-          "to" => to_address_hash,
-          "transactionIndex" => index,
-          "v" => v,
-          "value" => value,
-          "type" => type,
-          "executionNode" => execution_node_hash,
-          "requestRecord" => wrapped
-        } = transaction
-      ) do
-    result = %{
-      block_hash: block_hash,
-      block_number: block_number,
-      from_address_hash: from_address_hash,
-      gas: gas,
-      gas_price: gas_price,
-      hash: hash,
-      index: index,
-      input: input,
-      nonce: nonce,
-      r: r,
-      s: s,
-      to_address_hash: to_address_hash,
-      v: v,
-      value: value,
-      transaction_index: index,
-      type: type
-    }
-
-    # credo:disable-for-next-line
-    result =
-      if Application.get_env(:explorer, :chain_type) == "suave" do
-        Map.merge(result, %{
-          execution_node_hash: execution_node_hash,
-          wrapped_type: quantity_to_integer(Map.get(wrapped, "type")),
-          wrapped_nonce: quantity_to_integer(Map.get(wrapped, "nonce")),
-          wrapped_to_address_hash: Map.get(wrapped, "to"),
-          wrapped_gas: quantity_to_integer(Map.get(wrapped, "gas")),
-          wrapped_gas_price: quantity_to_integer(Map.get(wrapped, "gasPrice")),
-          wrapped_max_priority_fee_per_gas: quantity_to_integer(Map.get(wrapped, "maxPriorityFeePerGas")),
-          wrapped_max_fee_per_gas: quantity_to_integer(Map.get(wrapped, "maxFeePerGas")),
-          wrapped_value: quantity_to_integer(Map.get(wrapped, "value")),
-          wrapped_input: Map.get(wrapped, "input"),
-          wrapped_v: quantity_to_integer(Map.get(wrapped, "v")),
-          wrapped_r: quantity_to_integer(Map.get(wrapped, "r")),
-          wrapped_s: quantity_to_integer(Map.get(wrapped, "s")),
-          wrapped_hash: Map.get(wrapped, "hash")
-        })
-      else
-        result
-      end
-
-    put_if_present(transaction, result, [
-      {"creates", :created_contract_address_hash},
-      {"block_timestamp", :block_timestamp}
-    ])
-  end
-
-  def elixir_to_params(
+  def do_elixir_to_params(
         %{
           "blockHash" => block_hash,
           "blockNumber" => block_number,
@@ -454,7 +344,7 @@ defmodule EthereumJSONRPC.Transaction do
     ])
   end
 
-  def elixir_to_params(
+  def do_elixir_to_params(
         %{
           "blockHash" => block_hash,
           "blockNumber" => block_number,
@@ -494,6 +384,44 @@ defmodule EthereumJSONRPC.Transaction do
       {"creates", :created_contract_address_hash},
       {"block_timestamp", :block_timestamp}
     ])
+  end
+
+  defp chain_type_fields(params, elixir) do
+    case Application.get_env(:explorer, :chain_type) do
+      "ethereum" ->
+        put_if_present(elixir, params, [
+          {"blobVersionedHashes", :blob_versioned_hashes},
+          {"maxFeePerBlobGas", :max_fee_per_blob_gas}
+        ])
+
+      "suave" ->
+        wrapped = Map.get(elixir, "requestRecord")
+
+        if is_nil(wrapped) do
+          params
+        else
+          params
+          |> Map.merge(%{
+            execution_node_hash: Map.get(elixir, "executionNode"),
+            wrapped_type: quantity_to_integer(Map.get(wrapped, "type")),
+            wrapped_nonce: quantity_to_integer(Map.get(wrapped, "nonce")),
+            wrapped_to_address_hash: Map.get(wrapped, "to"),
+            wrapped_gas: quantity_to_integer(Map.get(wrapped, "gas")),
+            wrapped_gas_price: quantity_to_integer(Map.get(wrapped, "gasPrice")),
+            wrapped_max_priority_fee_per_gas: quantity_to_integer(Map.get(wrapped, "maxPriorityFeePerGas")),
+            wrapped_max_fee_per_gas: quantity_to_integer(Map.get(wrapped, "maxFeePerGas")),
+            wrapped_value: quantity_to_integer(Map.get(wrapped, "value")),
+            wrapped_input: Map.get(wrapped, "input"),
+            wrapped_v: quantity_to_integer(Map.get(wrapped, "v")),
+            wrapped_r: quantity_to_integer(Map.get(wrapped, "r")),
+            wrapped_s: quantity_to_integer(Map.get(wrapped, "s")),
+            wrapped_hash: Map.get(wrapped, "hash")
+          })
+        end
+
+      _ ->
+        params
+    end
   end
 
   @doc """
@@ -605,7 +533,7 @@ defmodule EthereumJSONRPC.Transaction do
   #
   # "txType": to avoid FunctionClauseError when indexing Wanchain
   defp entry_to_elixir({key, value})
-       when key in ~w(blockHash condition creates from hash input jsonrpc publicKey raw to txType executionNode requestRecord),
+       when key in ~w(blockHash condition creates from hash input jsonrpc publicKey raw to txType executionNode requestRecord blobVersionedHashes),
        do: {key, value}
 
   # specific to Nethermind client
@@ -613,19 +541,9 @@ defmodule EthereumJSONRPC.Transaction do
     do: {"input", value}
 
   defp entry_to_elixir({key, quantity})
-       when key in ~w(gas gasPrice nonce r s standardV v value type maxPriorityFeePerGas maxFeePerGas) and
+       when key in ~w(gas gasPrice nonce r s standardV v value type maxPriorityFeePerGas maxFeePerGas maxFeePerBlobGas) and
               quantity != nil do
     {key, quantity_to_integer(quantity)}
-  end
-
-  # to be merged with the clause above ^
-  defp entry_to_elixir({"maxFeePerBlobGas", _value}) do
-    {nil, nil}
-  end
-
-  # EIP-4844 specific field with value of type of list of hashes
-  defp entry_to_elixir({"blobVersionedHashes", _value}) do
-    {nil, nil}
   end
 
   # as always ganache has it's own vision on JSON RPC standard